Delete notification bug fix

+++ webapp/appModules/surveyForm/record/actions.js
@@ -119,16 +119,19 @@ export const removeNode = (nodeDef, node) => async (dispatch, getState) => {
   recordNodesUpdate(data.nodes)(dispatch)
 }
 
-export const deleteRecord = (history) => async (dispatch, getState) => {
-  const state = getState()
+export const deleteRecord = (history, server = true) => async (dispatch, getState) => {
+  if (server) {
+    const state = getState()
 
-  const surveyId = SurveyState.getStateSurveyId(state)
-  const recordUuid = RecordState.getRecordUuid(state)
+    const surveyId = SurveyState.getStateSurveyId(state)
+    const recordUuid = RecordState.getRecordUuid(state)
 
-  // 1. checkout (close server thread)
-  await checkOutRecord(recordUuid)(dispatch, getState)
-  // 2. perform server side delete
-  await axios.delete(`/api/survey/${surveyId}/record/${recordUuid}`)
+    // 1. checkout (close server thread)
+    await checkOutRecord(recordUuid)(dispatch, getState)
+    // 2. perform server side delete
+    await axios.delete(`/api/survey/${surveyId}/record/${recordUuid}`)
+  }
+
   // 3. remove record from redux state
   await dispatch({ type: recordDelete })
   // 4. redirect to default data module (records view)
